requirements.txt
fastapi~=0.110.1
sqlalchemy~=2.0.29
psycopg2-binary~=2.9.9
uvicorn~=0.29.0
alembic~=1.13.1
python-dotenv~=1.0.0
python-multipart~=0.0.9
python-jose~=3.3.0
cryptography~=42.0.5
passlib~=1.7.4
bcrypt==4.0.1
httpx
pytest